Changing the main functionality of the network, from forwarding votes, to forwarding requests for nones
