+++ .env.example
@@ -1,2 +1,2 @@
-node_id: 1
-next_node: http://127.0.0.1:8088/forward
+NODE_ID=1
+NEXT_NODE=http://127.0.0.1:8088/forward
