--- src/models/mod.rs
@@ -1,3 +1,0 @@
-mod vote;
-
-pub use vote::Vote;
