--- src/models/vote.rs
@@ -1,8 +1,0 @@
-use serde::{Serialize, Deserialize};
-
-#[derive(Serialize, Deserialize)]
-pub struct Vote {
-    pub election_id: u32,
-    pub voter_id: u32,
-    pub option: u32,
-}
